Remove support for 7.2
As PHP7.2 is by now not supported by the PHP-Project anymore I removed
support for it from this library.

Ich you are running into issues with this library under 7.2, update your
PHP Version to a supported one and check whether the bug still exists.
If so: Feel free to raise an issue