Add alternative forward and rewind info
That way we can implement rules where saturday and sunday observances
are forwarded to monday but wednesday and thursday observances are
forwarded to friday.

In addition to adding the rule to the XSD this also contains refactoring
to allow implementing these rules in an easier way for Date and Relative
implementations. Easter-based implementations don't need this as they
always fall on the same weekday and so there is no need to swap them.